Implemented ProtocolLayer_init() to initialize the Ethernet interface.
- Moved the implementation of protocol layer functions to protocol_layer.c.
- Ensured proper inclusion of necessary paths, headers and definitions.
- Resolved multiple definition errors by using extern declarations in protocol_layer.h.
- Implemented ProtocolLayer_init() to initialize the Ethernet interface.
