+++ source/enet_txrx_transfer.c
@@ -7,251 +7,34 @@
 
 #include "fsl_debug_console.h"
 #include "fsl_silicon_id.h"
-#include "fsl_enet.h"
-#include "fsl_phy.h"
+#include "protocol_layer.h"
 #include "board.h"
 #include "app.h"
 
 /*******************************************************************************
  * Definitions
  ******************************************************************************/
-#define ENET_RXBD_NUM          (4)
-#define ENET_TXBD_NUM          (4)
-#define ENET_RXBUFF_SIZE       (ENET_FRAME_MAX_FRAMELEN)
-#define ENET_TXBUFF_SIZE       (ENET_FRAME_MAX_FRAMELEN)
-#define ENET_DATA_LENGTH       (1000)
-#define ENET_TRANSMIT_DATA_NUM (20)
-#ifndef APP_ENET_BUFF_ALIGNMENT
-#define APP_ENET_BUFF_ALIGNMENT ENET_BUFF_ALIGNMENT
-#endif
-#ifndef PHY_AUTONEGO_TIMEOUT_COUNT
-#define PHY_AUTONEGO_TIMEOUT_COUNT (300000)
-#endif
-#ifndef EXAMPLE_PHY_LINK_INTR_SUPPORT
-#define EXAMPLE_PHY_LINK_INTR_SUPPORT (0U)
-#endif
-#ifndef EXAMPLE_USES_LOOPBACK_CABLE
-#define EXAMPLE_USES_LOOPBACK_CABLE (1U)
-#endif
-
-#ifndef PHY_STABILITY_DELAY_US
-#if EXAMPLE_USES_LOOPBACK_CABLE
-#define PHY_STABILITY_DELAY_US (0U)
-#else
-/* If cable is not used there is no "readiness wait" caused by auto negotiation. Lets wait 100ms.*/
-#define PHY_STABILITY_DELAY_US (100000U)
-#endif
-#endif
-
-/* @TEST_ANCHOR */
-
-#ifndef MAC_ADDRESS
-#define MAC_ADDRESS                        \
-    {                                      \
-        0x54, 0x27, 0x8d, 0x00, 0x00, 0x00 \
-    }
-#else
-#define USER_DEFINED_MAC_ADDRESS
-#endif
 
 /*******************************************************************************
  * Prototypes
  ******************************************************************************/
-/*! @brief Build ENET broadcast frame. */
-static void ENET_BuildBroadCastFrame(void);
-
-#if (defined(EXAMPLE_PHY_LINK_INTR_SUPPORT) && (EXAMPLE_PHY_LINK_INTR_SUPPORT))
-void GPIO_EnableLinkIntr(void);
-#endif
 
 /*******************************************************************************
  * Variables
  ******************************************************************************/
-/*! @brief Buffer descriptors should be in non-cacheable region and should be align to "ENET_BUFF_ALIGNMENT". */
-AT_NONCACHEABLE_SECTION_ALIGN(enet_rx_bd_struct_t g_rxBuffDescrip[ENET_RXBD_NUM], ENET_BUFF_ALIGNMENT);
-AT_NONCACHEABLE_SECTION_ALIGN(enet_tx_bd_struct_t g_txBuffDescrip[ENET_TXBD_NUM], ENET_BUFF_ALIGNMENT);
-/*! @brief The data buffers can be in cacheable region or in non-cacheable region.
- * If use cacheable region, the alignment size should be the maximum size of "CACHE LINE SIZE" and "ENET_BUFF_ALIGNMENT"
- * If use non-cache region, the alignment size is the "ENET_BUFF_ALIGNMENT".
- */
-SDK_ALIGN(uint8_t g_rxDataBuff[ENET_RXBD_NUM][SDK_SIZEALIGN(ENET_RXBUFF_SIZE, APP_ENET_BUFF_ALIGNMENT)],
-          APP_ENET_BUFF_ALIGNMENT);
-SDK_ALIGN(uint8_t g_txDataBuff[ENET_TXBD_NUM][SDK_SIZEALIGN(ENET_TXBUFF_SIZE, APP_ENET_BUFF_ALIGNMENT)],
-          APP_ENET_BUFF_ALIGNMENT);
-
-/*! @brief MAC transfer. */
-static enet_handle_t g_handle;
-static uint8_t g_frame[ENET_DATA_LENGTH + 14];
-
-/*! @brief The MAC address for ENET device. */
-uint8_t g_macAddr[6] = MAC_ADDRESS;
-
-/*! @brief PHY status. */
-static phy_handle_t phyHandle;
-#if ((EXAMPLE_USES_LOOPBACK_CABLE) && defined(EXAMPLE_PHY_LINK_INTR_SUPPORT) && (EXAMPLE_PHY_LINK_INTR_SUPPORT))
-static bool linkChange = false;
-#endif
 
 /*******************************************************************************
  * Code
  ******************************************************************************/
-/*! @brief Build Frame for transmit. */
-static void ENET_BuildBroadCastFrame(void)
-{
-    uint32_t count  = 0;
-    uint32_t length = ENET_DATA_LENGTH - 14;
-
-    for (count = 0; count < 6U; count++)
-    {
-        g_frame[count] = 0xFFU;
-    }
-    memcpy(&g_frame[6], &g_macAddr[0], 6U);
-    g_frame[12] = (length >> 8) & 0xFFU;
-    g_frame[13] = length & 0xFFU;
-
-    for (count = 0; count < length; count++)
-    {
-        g_frame[count + 14] = count % 0xFFU;
-    }
-}
-
-#if (defined(EXAMPLE_PHY_LINK_INTR_SUPPORT) && (EXAMPLE_PHY_LINK_INTR_SUPPORT))
-void PHY_LinkStatusChange(void)
-{
-#if (EXAMPLE_USES_LOOPBACK_CABLE)
-    linkChange = true;
-#endif
-}
-#endif
 
 /*!
  * @brief Main function
  */
 int main(void)
 {
-    phy_config_t phyConfig = {0};
-    uint32_t testTxNum     = 0;
-    uint32_t length        = 0;
-    enet_data_error_stats_t eErrStatic;
-    status_t status;
-    enet_config_t config;
-#if EXAMPLE_USES_LOOPBACK_CABLE
-    volatile uint32_t count = 0;
-    phy_speed_t speed;
-    phy_duplex_t duplex;
-    bool autonego = false;
-    bool link     = false;
-    bool tempLink = false;
-#endif
-
     /* Hardware Initialization. */
     BOARD_InitHardware();
-
-    PRINTF("\r\nENET example start.\r\n");
-
-    /* Prepare the buffer configuration. */
-    enet_buffer_config_t buffConfig[] = {{
-        ENET_RXBD_NUM,
-        ENET_TXBD_NUM,
-        SDK_SIZEALIGN(ENET_RXBUFF_SIZE, APP_ENET_BUFF_ALIGNMENT),
-        SDK_SIZEALIGN(ENET_TXBUFF_SIZE, APP_ENET_BUFF_ALIGNMENT),
-        &g_rxBuffDescrip[0],
-        &g_txBuffDescrip[0],
-        &g_rxDataBuff[0][0],
-        &g_txDataBuff[0][0],
-        true,
-        true,
-        NULL,
-    }};
-
-    /* Get default configuration. */
-    /*
-     * config.miiMode = kENET_RmiiMode;
-     * config.miiSpeed = kENET_MiiSpeed100M;
-     * config.miiDuplex = kENET_MiiFullDuplex;
-     * config.rxMaxFrameLen = ENET_FRAME_MAX_FRAMELEN;
-     */
-    ENET_GetDefaultConfig(&config);
-
-    /* The miiMode should be set according to the different PHY interfaces. */
-#ifdef EXAMPLE_PHY_INTERFACE_RGMII
-    config.miiMode = kENET_RgmiiMode;
-#else
-    config.miiMode = kENET_RmiiMode;
-#endif
-    phyConfig.phyAddr = EXAMPLE_PHY_ADDRESS;
-#if EXAMPLE_USES_LOOPBACK_CABLE
-    phyConfig.autoNeg = true;
-#else
-    phyConfig.autoNeg = false;
-    config.miiDuplex  = kENET_MiiFullDuplex;
-#endif
-    phyConfig.ops      = EXAMPLE_PHY_OPS;
-    phyConfig.resource = EXAMPLE_PHY_RESOURCE;
-#if (defined(EXAMPLE_PHY_LINK_INTR_SUPPORT) && (EXAMPLE_PHY_LINK_INTR_SUPPORT))
-    phyConfig.intrType = kPHY_IntrActiveLow;
-#endif
-
-    /* Initialize PHY and wait auto-negotiation over. */
-    PRINTF("Wait for PHY init...\r\n");
-#if EXAMPLE_USES_LOOPBACK_CABLE
-    do
-    {
-        status = PHY_Init(&phyHandle, &phyConfig);
-        if (status == kStatus_Success)
-        {
-            PRINTF("Wait for PHY link up...\r\n");
-            /* Wait for auto-negotiation success and link up */
-            count = PHY_AUTONEGO_TIMEOUT_COUNT;
-            do
-            {
-                PHY_GetLinkStatus(&phyHandle, &link);
-                if (link)
-                {
-                    PHY_GetAutoNegotiationStatus(&phyHandle, &autonego);
-                    if (autonego)
-                    {
-                        break;
-                    }
-                }
-            } while (--count);
-            if (!autonego)
-            {
-                PRINTF("PHY Auto-negotiation failed. Please check the cable connection and link partner setting.\r\n");
-            }
-        }
-    } while (!(link && autonego));
-#else
-    while (PHY_Init(&phyHandle, &phyConfig) != kStatus_Success)
-    {
-        PRINTF("PHY_Init failed\r\n");
-    }
-
-    /* set PHY link speed/duplex and enable loopback. */
-    PHY_SetLinkSpeedDuplex(&phyHandle, (phy_speed_t)config.miiSpeed, (phy_duplex_t)config.miiDuplex);
-    PHY_EnableLoopback(&phyHandle, kPHY_LocalLoop, (phy_speed_t)config.miiSpeed, true);
-#endif /* EXAMPLE_USES_LOOPBACK_CABLE */
-
-#if PHY_STABILITY_DELAY_US
-    /* Wait a moment for PHY status to be stable. */
-    SDK_DelayAtLeastUs(PHY_STABILITY_DELAY_US, SDK_DEVICE_MAXIMUM_CPU_CLOCK_FREQUENCY);
-#endif
-
-#if EXAMPLE_USES_LOOPBACK_CABLE
-    /* Get the actual PHY link speed and set in MAC. */
-    PHY_GetLinkSpeedDuplex(&phyHandle, &speed, &duplex);
-    config.miiSpeed  = (enet_mii_speed_t)speed;
-    config.miiDuplex = (enet_mii_duplex_t)duplex;
-#endif
-
-#ifndef USER_DEFINED_MAC_ADDRESS
-    /* Set special address for each chip. */
-    SILICONID_ConvertToMacAddr(&g_macAddr);
-#endif
-
-    /* Init the ENET. */
-    ENET_Init(EXAMPLE_ENET, &g_handle, &config, &buffConfig[0], &g_macAddr[0], EXAMPLE_CLOCK_FREQ);
-    ENET_ActiveRead(EXAMPLE_ENET);
+    ProtocolLayer_init();
 
     /* Build broadcast for sending. */
     ENET_BuildBroadCastFrame();
